refactor: split llm_providers.py into src/llm/ package (6 modules)
