+++ src/llm/prompt_builders.py
@@ -1,0 +1,390 @@
+"""Unified prompt builders for all LLM providers."""
+from typing import Dict, Any, Optional, List
+from loguru import logger
+from src.config import settings
+
+
+def _build_system_prompt(
+    transcription: Optional[str] = None,
+    diarization_analysis: Optional[Dict[str, Any]] = None,
+    llm_meeting_type: Optional[str] = None
+) -> str:
+    """
+    Strict system policy for generating professional protocol.
+    Automatically selects specialized prompt if classification is enabled.
+    """
+    meeting_type = None
+
+    if llm_meeting_type:
+        meeting_type = llm_meeting_type
+        logger.info(f"Использую тип встречи определенный LLM: {meeting_type}")
+    elif settings.meeting_type_detection and transcription:
+        try:
+            from src.services.meeting_classifier import meeting_classifier
+            meeting_type, _ = meeting_classifier.classify(
+                transcription,
+                diarization_analysis
+            )
+            logger.info(f"Использую тип встречи определенный классификатором: {meeting_type}")
+        except Exception as e:
+            logger.warning(f"Ошибка при классификации встречи: {e}. Используем базовый промпт")
+
+    if meeting_type:
+        logger.info(f"Определен тип встречи: {meeting_type} (используем базовый промпт)")
+
+    base_prompt = (
+        "Ты — профессиональный протоколист высшей квалификации с опытом документирования "
+        "деловых встреч, совещаний и переговоров.\n\n"
+
+        "ТВОЯ РОЛЬ:\n"
+        "- Извлекать и структурировать ключевую информацию из стенограмм встреч\n"
+        "- Создавать четкие, лаконичные и информативные протоколы\n"
+        "- Сохранять объективность и фактологическую точность\n\n"
+
+        "ФУНДАМЕНТАЛЬНЫЕ ПРИНЦИПЫ ТОЧНОСТИ:\n"
+        "- АБСОЛЮТНАЯ ПОЛНОТА: Анализируй ВСЮ транскрипцию от первого до последнего слова. "
+        "Каждая упомянутая тема, идея, вопрос или комментарий должны быть отражены в протоколе\n"
+        "- ФАКТИЧЕСКАЯ ТОЧНОСТЬ: Используй ТОЛЬКО информацию, явно присутствующую в транскрипции. "
+        "НЕ добавляй, НЕ домысливай, НЕ интерпретируй сверх сказанного\n"
+        "- СОХРАНЕНИЕ КОНТЕКСТА: Фиксируй не только ЧТО было сказано, но и КАК обсуждалось "
+        "(тон дискуссии, уровень согласия/несогласия, степень проработки вопроса)\n"
+        "- ДЕТАЛИЗАЦИЯ БЕЗ ПОТЕРЬ: Если участник упомянул конкретный пример, цифру, дату, имя, "
+        "технологию или любую специфическую деталь — это ОБЯЗАТЕЛЬНО должно быть в протоколе\n"
+        "- АТРИБУЦИЯ ВЫСКАЗЫВАНИЙ: Если в транскрипции указаны имена участников, ВСЕГДА указывай, "
+        "кто что сказал (например: \"Иван предложил...\", \"Мария отметила...\")\n"
+        "- ОТКРЫТЫЕ ВОПРОСЫ: Фиксируй ВСЕ незакрытые вопросы, темы, требующие дополнительного обсуждения, "
+        "и моменты неопределенности\n\n"
+
+        "ПРАВИЛА ОБРАБОТКИ СПЕЦИФИЧЕСКОЙ ИНФОРМАЦИИ:\n"
+        "- Ответственные лица: Указывай ТОЛЬКО если явно названы в транскрипции. "
+        "Если сказано \"кто-то должен это сделать\" без имени — пиши \"не определен\" или \"Не указано\"\n"
+        "- Сроки и даты: Указывай ТОЛЬКО если явно упомянуты. Если сказано \"скоро\" или \"в ближайшее время\" — "
+        "сохраняй эту формулировку\n"
+        "- Решения: Четко различай между \"принято решение\" и \"обсуждалась возможность\". "
+        "Фиксируй уровень определенности\n"
+        "- Цифры и метрики: Записывай все упомянутые числа, проценты, бюджеты, сроки точно как в транскрипции\n\n"
+
+        "ПРИНЦИПЫ РАБОТЫ:\n"
+        "1. ТОЧНОСТЬ: Используй только факты, явно присутствующие в стенограмме\n"
+        "2. НЕТ ДОМЫСЛОВ: Не додумывай, не интерпретируй, не добавляй информацию от себя\n"
+        "3. КОНТЕКСТ: Если упоминается роль/должность/срок/сумма — укажи их; если нет — не придумывай\n"
+        "4. КРАТКОСТЬ: Излагай суть без воды, избегай избыточных деталей\n"
+        "5. ТЕРМИНОЛОГИЯ: Сохраняй профессиональные термины и названия как в оригинале\n"
+        "6. СТИЛЬ: Официально-деловой язык без разговорных оборотов\n\n"
+
+        "КРИТИЧЕСКИ ВАЖНО - ФОРМАТ ИМЕН УЧАСТНИКОВ В ПРОТОКОЛЕ:\n"
+        "В секции 'Участники' протокола используй имена в формате 'Имя Фамилия' БЕЗ отчества!\n"
+        "Если предоставлен список участников - КОПИРУЙ имена ТОЧНО как они указаны в списке.\n\n"
+        "НЕПРАВИЛЬНО: только имя ('Софья', 'Галина') или с отчеством ('Софья Юрьевна Осипова')\n"
+        "ПРАВИЛЬНО: 'Софья Осипова', 'Галина Ямкина', 'Владимир Голиков'\n\n"
+        "ОПРЕДЕЛЕНИЕ УЧАСТНИКОВ ИЗ ТРАНСКРИПЦИИ:\n"
+        "Если список участников не предоставлен явно:\n"
+        "- Извлекай имена ТОЛЬКО из явных упоминаний в транскрипции\n"
+        "- Источники: представления ('Меня зовут...'), обращения ('Света, как думаешь?')\n"
+        "- Преобразуй уменьшительные в полные: Света->Светлана, Леша->Алексей\n"
+        "- Используй формат 'Имя Фамилия' где возможно\n"
+        "- ЕСЛИ имя не определить - оставляй метку спикера (SPEAKER_1, SPEAKER_2)\n"
+        "- НЕ придумывай имена! НЕ используй 'Участник 1', 'Коллега'\n"
+        "- НЕ дублируй: если Света = SPEAKER_1, не добавляй Светлану отдельно\n\n"
+
+        "ФОРМАТИРОВАНИЕ ОБСУЖДЕНИЯ:\n"
+        "Если группируешь обсуждение по темам/кластерам:\n"
+        "- НЕ пиши слово 'Кластер', только название темы с маркером: '* **Название темы**'\n"
+        "- Каждую идею/высказывание/позицию с новой строки\n"
+        "- Формат высказывания: 'Имя Автора: текст' (без слова 'Идея', без скобок)\n"
+        "- Между тематическими блоками оставляй пустую строку для визуального разделения\n\n"
+        "Примеры:\n"
+        "ПРАВИЛЬНО:\n"
+        "* **Архитектурные решения**\n\n"
+        "Алексей Тимченко: предложил использовать микросервисную архитектуру\n\n"
+        "Мария Иванова: поддержала, добавила про важность API gateway\n\n"
+        "НЕПРАВИЛЬНО:\n"
+        "* Кластер 'Архитектурные решения': Идея (Алексей Тимченко): предложил...\n\n"
+
+        "ЧТО ИГНОРИРОВАТЬ:\n"
+        "- Междометия (э-э, м-м, ну, вот)\n"
+        "- Повторы и запинки\n"
+        "- Вводные фразы без смысловой нагрузки\n"
+        "- Отвлеченные разговоры не по теме встречи\n\n"
+
+        "ЧТО ВЫДЕЛЯТЬ:\n"
+        "- Конкретные решения и резолюции\n"
+        "- Поручения с указанием исполнителей (если упомянуты)\n"
+        "- Сроки, суммы, показатели (только если явно названы)\n"
+        "- Ключевые проблемы и их решения\n"
+        "- Согласованные договоренности\n"
+        "- Цифры, метрики, конкретные примеры (ОБЯЗАТЕЛЬНО включай все упомянутые)\n"
+        "- Все темы обсуждения с указанием кто что сказал\n"
+        "- Моменты несогласия, альтернативные мнения, нерешенные вопросы\n\n"
+
+        "ФОРМАТ ВЫВОДА:\n"
+        "Верни JSON в формате схемы. Если данные отсутствуют или неоднозначны — используй 'Не указано'.\n\n"
+    )
+
+    formatting_rules = (
+        "КРИТИЧЕСКИ ВАЖНО — форматирование значений полей протокола:\n"
+        "- ВСЕ значения полей должны быть ПРОСТЫМИ СТРОКАМИ (string), не объектами или массивами\n"
+        "- НЕ используй вложенные объекты {} или массивы [] в качестве значений полей протокола\n"
+        "- Списки форматируй как многострочный текст с маркерами '- ' (дефис + пробел)\n"
+        "- Даты и время: простой текст, например '20 октября 2024, 14:30'\n"
+        "- Участники: каждое имя с новой строки через \\n, БЕЗ ролей!, например 'Иван Петров\\nМария Сидорова\\nАлексей Смирнов'\n"
+        "- Решения и задачи: многострочный текст со списком через \\n, каждый пункт с '- '\n\n"
+
+        "ФОРМАТИРОВАНИЕ СПИСКОВ:\n"
+        "- Внутри значений JSON полей: используй одинарный \\n для разделения элементов списка\n"
+        "- Для визуального разделения тематических блоков в обсуждении: используй двойной \\n\\n\n\n"
+
+        "ПРИМЕР ПРАВИЛЬНОГО JSON:\n"
+        "{\n"
+        '  "date": "20 октября 2024",\n'
+        '  "time": "14:30",\n'
+        '  "participants": "Оксана Иванова\\nГалина Петрова\\nАлексей Смирнов",\n'
+        '  "decisions": "- Решение 1\\n- Решение 2\\n- Решение 3"\n'
+        "}\n\n"
+
+        "ПРИМЕР НЕПРАВИЛЬНОГО JSON (НЕ ДЕЛАЙ ТАК):\n"
+        "{\n"
+        '  "date": {"day": 20, "month": "октябрь"},  вложенный объект\n'
+        '  "participants": ["Оксана", "Галя"],  массив\n'
+        '  "decisions": [{"decision": "Решение 1"}]  массив объектов\n'
+        "}"
+    )
+
+    return base_prompt + formatting_rules
+
+
+def _build_user_prompt(
+    transcription: str,
+    template_variables: Dict[str, str],
+    diarization_data: Optional[Dict[str, Any]] = None,
+    speaker_mapping: Optional[Dict[str, str]] = None,
+    meeting_topic: Optional[str] = None,
+    meeting_date: Optional[str] = None,
+    meeting_time: Optional[str] = None,
+    participants: Optional[List[Dict[str, str]]] = None,
+    meeting_agenda: Optional[str] = None,
+    project_list: Optional[str] = None,
+) -> str:
+    """
+    Build user prompt with context and format requirements.
+    """
+    from src.prompts.prompts import _build_field_specific_rules
+
+    # Transcription block (with diarization support)
+    if diarization_data and diarization_data.get("formatted_transcript"):
+        transcription_text = (
+            "Транскрипция с разделением говорящих:\n"
+            f"{diarization_data['formatted_transcript']}\n\n"
+            "Дополнительная информация:\n"
+            f"- Количество говорящих: {diarization_data.get('total_speakers', 'неизвестно')}\n"
+            f"- Список говорящих: {', '.join(diarization_data.get('speakers', []))}\n\n"
+        )
+        if settings.include_raw_transcription_in_prompts:
+            transcription_text += (
+                "\n" + "=" * 53 + "\n"
+                "ИСХОДНАЯ ТРАНСКРИПЦИЯ (ДЛЯ ОТЛАДКИ):\n"
+                "=" * 53 + "\n\n"
+                f"{transcription}\n\n"
+            )
+    else:
+        transcription_text = (
+            "Транскрипция:\n"
+            f"{transcription}\n\n"
+            "Примечание: Диаризация (разделение говорящих) недоступна для этой записи.\n"
+        )
+
+    # Participants info
+    participants_info = ""
+    if speaker_mapping:
+        mapping_list = "\n".join(f"- {sid} = {name}" for sid, name in speaker_mapping.items())
+        participants_info = (
+            f"\n\n{'=' * 63}\n"
+            "УЧАСТНИКИ ВСТРЕЧИ (С РОЛЯМИ)\n"
+            f"{'=' * 63}\n\n"
+            "Сопоставление говорящих с участниками:\n"
+            f"{mapping_list}\n\n"
+            "ИНСТРУКЦИИ ПО РАБОТЕ С УЧАСТНИКАМИ:\n"
+            "1. Используй РЕАЛЬНЫЕ ИМЕНА вместо меток спикеров (SPEAKER_1 -> Имя)\n"
+            "2. При назначении ответственных учитывай контекст высказываний участников\n"
+            "3. Формат ответственного: ТОЛЬКО ИМЯ, без роли в скобках\n"
+            "   Правильно: 'Ответственный: Иван Петров'\n"
+            "   Неправильно: 'Ответственный: Иван Петров (Менеджер)'\n"
+            "СОПОСТАВЛЕНИЕ ИМЕН:\n"
+            "В транскрипции могут встречаться сокращенные/разговорные варианты имен.\n"
+            "АВТОМАТИЧЕСКИ сопоставляй их с полными именами из списка выше:\n\n"
+            "Примеры логики сопоставления (применяй ко ВСЕМ участникам):\n"
+            "   Уменьшительные: Света->Светлана, Леша->Алексей, Саша->Александр и т.д.\n"
+            "   По фамилии: Тимченко->Алексей Тимченко, Короткова->Светлана Короткова и т.д.\n"
+            "   Только имя: Алексей->Алексей Тимченко (если один такой в списке)\n\n"
+            "   НЕ ограничивайся примерами! Анализируй ВЕСЬ список участников выше.\n"
+            "   В финальном протоколе используй ПОЛНОЕ ИМЯ из списка участников!\n"
+        )
+    elif participants:
+        from src.services.participants_service import participants_service
+        formatted_participants = participants_service.format_participants_for_llm(participants)
+
+        participants_info = (
+            f"\n\n{'=' * 63}\n"
+            "ПОЛНЫЙ СПИСОК УЧАСТНИКОВ ВСТРЕЧИ (ОБЯЗАТЕЛЬНЫЙ К ИСПОЛЬЗОВАНИЮ)\n"
+            f"{'=' * 63}\n\n"
+            f"{formatted_participants}\n\n"
+            "КРИТИЧЕСКИ ВАЖНО - СТРОГИЕ ПРАВИЛА ИСПОЛЬЗОВАНИЯ\n\n"
+            "1. ИСПОЛЬЗУЙ ТОЛЬКО ИМЕНА ИЗ СПИСКА ВЫШЕ!\n"
+            "   ЗАПРЕЩЕНО добавлять участников, которых НЕТ в списке!\n\n"
+            "2. ФОРМАТ ИМЕН: 'Имя Фамилия' (БЕЗ отчества)!\n\n"
+            "3. СОПОСТАВЛЕНИЕ СОКРАЩЕННЫХ ИМЕН:\n"
+            "   В транскрипции упоминания могут быть сокращенными.\n"
+            "   ОБЯЗАТЕЛЬНО найди в списке выше ПОЛНОЕ соответствие.\n\n"
+            "4. ПРОВЕРКА ПЕРЕД ДОБАВЛЕНИЕМ В ПРОТОКОЛ:\n"
+            "   Для КАЖДОГО участника из транскрипции:\n"
+            "   - Найди соответствие в списке выше\n"
+            "   - Используй ТОЧНОЕ написание из списка (Имя Фамилия)\n"
+            "   - Если не можешь определить конкретного человека - НЕ включай в протокол\n\n"
+            "ВАЖНО: Проанализируй ВЕСЬ список участников выше!\n"
+            "НЕ выдумывай имена! Используй ТОЛЬКО из списка!\n\n"
+        )
+    else:
+        participants_info = (
+            f"\n\n{'=' * 63}\n"
+            "АВТОМАТИЧЕСКОЕ ОПРЕДЕЛЕНИЕ УЧАСТНИКОВ ИЗ ТРАНСКРИПЦИИ\n"
+            f"{'=' * 63}\n\n"
+            "Список участников не предоставлен. Определи имена из транскрипции.\n\n"
+            "ПРАВИЛА ОПРЕДЕЛЕНИЯ:\n\n"
+            "1. ИЩИ ЯВНЫЕ УПОМИНАНИЯ:\n"
+            "   Представления: 'Меня зовут Иван Петров', 'Я - Мария'\n"
+            "   Обращения: 'Света, как думаешь?', 'Петров, расскажи о задаче'\n"
+            "   Упоминания: 'Как сказал Иван...', 'Нужно уточнить у Марии'\n\n"
+            "2. ФОРМАТ ИМЕН:\n"
+            "   Предпочтительно: 'Имя Фамилия' (БЕЗ отчества)\n"
+            "   Если известно только имя: 'Иван'\n"
+            "   Если известна только фамилия: 'Петров'\n"
+            "   Преобразуй уменьшительные: Света->Светлана, Леша->Алексей, Володя->Владимир\n\n"
+            "3. СОПОСТАВЛЕНИЕ СО СПИКЕРАМИ:\n"
+            "   Сопоставь каждую метку (SPEAKER_1, SPEAKER_2...) с именем если возможно\n"
+            "   Если имя определить НЕВОЗМОЖНО - оставь метку спикера как есть\n\n"
+            "4. СТРОГИЕ ЗАПРЕТЫ:\n"
+            "   НЕ придумывай имена, которых НЕТ в транскрипции\n"
+            "   НЕ используй 'Участник 1', 'Коллега', 'Человек А', 'Неизвестный'\n"
+            "   НЕ дублируй: если Света = SPEAKER_1, не добавляй Светлану отдельно\n"
+            "   НЕ заменяй SPEAKER_N на описания типа 'Руководитель встречи'\n\n"
+        )
+
+    # Meeting info
+    meeting_info = ""
+    if meeting_topic or meeting_date or meeting_time:
+        meeting_info = "\n\n" + "=" * 63 + "\n"
+        meeting_info += "ИНФОРМАЦИЯ О ВСТРЕЧЕ\n"
+        meeting_info += "=" * 63 + "\n\n"
+        if meeting_topic:
+            meeting_info += f"Тема: {meeting_topic}\n"
+        if meeting_date:
+            meeting_info += f"Дата: {meeting_date}\n"
+        if meeting_time:
+            meeting_info += f"Время: {meeting_time}\n"
+        meeting_info += "\n"
+
+    # Additional context section (agenda, projects)
+    context_section = ""
+    if meeting_agenda or project_list:
+        context_info = {
+            'meeting_agenda': meeting_agenda,
+            'project_list': project_list
+        }
+        context_info = {k: v for k, v in context_info.items() if v}
+
+        if context_info:
+            context_section += "\n\n## ДОПОЛНИТЕЛЬНЫЙ КОНТЕКСТ ДЛЯ АНАЛИЗА\n\n"
+            context_section += "Используй следующую информацию для более точного извлечения данных протокола:\n\n"
+
+            for key, value in context_info.items():
+                if key == 'meeting_agenda':
+                    context_section += f"**Повестка встречи:**\n{value}\n\n"
+                elif key == 'project_list':
+                    context_section += f"**Список проектов:**\n{value}\n\n"
+
+    variables_str = "\n".join([f"- {key}: {desc}" for key, desc in template_variables.items()])
+
+    user_prompt = (
+        "=" * 59 + "\n"
+        "ИСХОДНЫЕ ДАННЫЕ ДЛЯ АНАЛИЗА\n"
+        "=" * 59 + "\n\n"
+        f"{context_section}"
+        f"{transcription_text}\n"
+        f"{participants_info}"
+        f"{meeting_info}"
+        "=" * 59 + "\n"
+        "ПОЛЯ ДЛЯ ИЗВЛЕЧЕНИЯ\n"
+        "=" * 59 + "\n\n"
+        f"{variables_str}\n\n"
+        "=" * 59 + "\n"
+        "ИНСТРУКЦИИ ПО ИЗВЛЕЧЕНИЮ\n"
+        "=" * 59 + "\n\n"
+
+        "СТРУКТУРА ВЫВОДА:\n"
+        "- Верни только валидный JSON-объект (без ```json, без markdown)\n"
+        "- Используй СТРОГО эти ключи из списка полей\n"
+        "- Сохраняй порядок ключей как в списке выше\n"
+        "- Каждое значение — строка (UTF-8), БЕЗ вложенных объектов или массивов\n\n"
+
+        "ФОРМАТИРОВАНИЕ ТЕКСТА:\n"
+        "- Для списков/перечислений: каждый пункт с новой строки, начинай с '- ' (дефис + пробел)\n"
+        "- НЕ используй нумерацию (1. 2. 3.), только дефисы\n"
+        "- НЕ ставь точку в конце пункта списка\n"
+        "- Для имен/участников: разделяй точкой с запятой (;)\n"
+        "- Для дат/времени: сохраняй формат как упомянут в тексте\n\n"
+
+        "ИЗВЛЕЧЕНИЕ ДАННЫХ:\n"
+        "- Используй ТОЛЬКО факты из транскрипции\n"
+        "- Если информация упомянута с контекстом (роль, срок, сумма) — укажи полностью\n"
+        "- Если данные отсутствуют, неоднозначны или неясны — пиши 'Не указано'\n"
+        "- Убирай дубликаты, объединяй идентичные пункты\n"
+        "- Сохраняй хронологию: порядок пунктов = порядок в тексте\n\n"
+
+        "ОБРАБОТКА ИНФОРМАЦИИ О ГОВОРЯЩИХ:\n"
+        "- Если доступна диаризация: используй метки 'Спикер 1:', 'Спикер 2:' и т.д.\n"
+        "- Определяй ответственных за действия по контексту их высказываний\n"
+        "- Указывай кто принял решение или взял на себя обязательство\n"
+        "- Если известны имена участников — используй ИХ, а не метки спикеров\n"
+        "- Формат задачи: 'Описание задачи — Ответственный: Имя Фамилия' (без роли!)\n"
+        "- Формат решения: 'Решение. Инициатор: Имя Фамилия' (если важно кто принял)\n\n"
+
+        "ЧТО ОТФИЛЬТРОВЫВАТЬ:\n"
+        "- Междометия, запинки, повторы слов\n"
+        "- Вводные фразы без смысловой нагрузки\n"
+        "- Разговоры не по теме встречи\n"
+        "- Технические комментарии ('не слышно', 'повторите' и т.д.)\n\n"
+
+        "=" * 59 + "\n"
+        "ПРИМЕРЫ ПРАВИЛЬНОГО ФОРМАТИРОВАНИЯ\n"
+        "=" * 59 + "\n\n"
+
+        "ПРАВИЛЬНО:\n"
+        "{\n"
+        "  \"participants\": \"Иван Иванов\\nМария Петрова\\nАлексей Сидоров\",\n"
+        "  \"main_topic\": \"Планирование маркетинговой кампании на Q2 2024\",\n"
+        "  \"decisions\": \"- Увеличить бюджет на digital-маркетинг на 30%\\n- Утвердить новую стратегию продвижения в социальных сетях\\n- Отложить запуск рекламы на ТВ до следующего квартала\",\n"
+        "  \"action_items\": \"- Подготовить презентацию новой стратегии к 15 марта — Ответственный: Мария Петрова\\n- Согласовать бюджет с финансовым отделом — Ответственный: Иван Иванов\\n- Провести анализ конкурентов — Ответственный: Алексей Сидоров\",\n"
+        "  \"deadlines\": \"- Презентация стратегии: 15 марта 2024\\n- Согласование бюджета: до конца текущей недели\\n- Анализ конкурентов: к следующему совещанию\",\n"
+        "  \"issues\": \"- Недостаточный охват целевой аудитории текущими каналами\\n- Высокая стоимость привлечения клиента\\n- Необходимость обновления креативов\"\n"
+        "}\n\n"
+        "ОБРАТИТЕ ВНИМАНИЕ: В action_items указаны только ИМЕНА без ролей в скобках!\n\n"
+
+        "НЕПРАВИЛЬНО:\n"
+        "{\n"
+        "  \"participants\": [\"Иван\", \"Мария\"],  массив вместо строки\n"
+        "  \"decisions\": \"1) Решение один 2) Решение два.\",  нумерация + точки\n"
+        "  \"action_items\": \"Подготовить презентацию (Мария)\",  без дефиса, неправильный формат\n"
+        "  \"deadlines\": \"Срочно, как можно быстрее\",  нет конкретики, хотя она могла быть\n"
+        "  \"extra_field\": \"...\",  поле не из списка\n"
+        "  \"budget\": \"50000 рублей (предположительно)\"  домыслы в скобках\n"
+        "}\n\n"
+
+        "=" * 59 + "\n"
+        "СПЕЦИФИЧНЫЕ ПРАВИЛА ПО ТИПАМ ПОЛЕЙ\n"
+        "=" * 59 + "\n\n"
+        f"{_build_field_specific_rules(template_variables)}\n\n"
+        "=" * 59 + "\n\n"
+
+        "НАЧИНАЙ АНАЛИЗ. Верни только JSON без дополнительных комментариев.\n"
+    )
+    return user_prompt
